+++ src/misc/httpError.interface.ts
@@ -65,3 +65,19 @@ export interface Handle400WithValidationsOptions extends Handle4xxOptions
      */
     clientValidationErrorsResponseMapper?: HttpClientValidationErrorResponseMapper;
 }
+
+/**
+ * Definition of http client errors read from http error response
+ */
+export interface HttpClientErrors
+{
+    /**
+     * Array of error
+     */
+    errors: string[];
+
+    /**
+     * Object storing validation errors
+     */
+    validationErrors: HttpClientValidationErrors|null;
+}